+++ CleanyModal/Classes/CleanyModalViewController.swift
@@ -18,8 +18,8 @@ open class CleanyModalViewController: UIViewController {
     
     @IBOutlet public var alertView: UIView!
     @IBOutlet public var alertViewCenterY: NSLayoutConstraint!
-    @IBOutlet public var widthConstraint: NSLayoutConstraint!
-
+    @IBOutlet public var alertWidthConstraint: NSLayoutConstraint!
+    @IBOutlet public var alertViewBottom: NSLayoutConstraint?
     
     public override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: Bundle?) {
         super.init(nibName: nibNameOrNil, bundle: nibBundleOrNil)