+++ CleanyModal/Classes/Helper/NSLayoutConstraint+Extension.swift
@@ -1,0 +1,66 @@
+//
+//  CleanyConfig.swift
+//  CleanyModal
+//
+//  Created by Lory Huz on 21/03/2018.
+//
+
+import UIKit
+
+extension NSLayoutConstraint {
+    
+    /**
+     Change priority constraint
+     
+     - parameter priority: CGFloat
+     - returns: NSLayoutConstraint
+     */
+    func setSafelyPriority(_ priority: UILayoutPriority) -> NSLayoutConstraint {
+        
+        NSLayoutConstraint.deactivate([self])
+        let newConstraint = NSLayoutConstraint(
+            item: firstItem ?? self,
+            attribute: firstAttribute,
+            relatedBy: relation,
+            toItem: secondItem,
+            attribute: secondAttribute,
+            multiplier: multiplier,
+            constant: constant)
+        
+        newConstraint.priority = priority
+        newConstraint.shouldBeArchived = self.shouldBeArchived
+        newConstraint.identifier = self.identifier
+        
+        NSLayoutConstraint.activate([newConstraint])
+        return newConstraint
+    }
+    
+    /**
+     Change multiplier constraint
+     
+     - parameter multiplier: CGFloat
+     - returns: NSLayoutConstraint
+     */
+    func setSafelyMultiplier(multiplier: CGFloat) -> NSLayoutConstraint {
+        
+        NSLayoutConstraint.deactivate([self])
+        let newConstraint = NSLayoutConstraint(
+            item: firstItem ?? self,
+            attribute: firstAttribute,
+            relatedBy: relation,
+            toItem: secondItem,
+            attribute: secondAttribute,
+            multiplier: multiplier,
+            constant: constant)
+        
+        newConstraint.priority = priority
+        newConstraint.shouldBeArchived = self.shouldBeArchived
+        newConstraint.identifier = self.identifier
+        
+        NSLayoutConstraint.activate([newConstraint])
+        return newConstraint
+    }
+    
+}
+
+