+++ CleanyModal/Classes/Model/AlertModel.swift
@@ -1,0 +1,16 @@
+//
+//  AlertModel.swift
+//  CleanyModal
+//
+//  Created by Lory Huz on 07/01/2020.
+//
+
+import Foundation
+
+struct AlertModel: Codable {
+    
+    var title: String?
+    var message: String?
+    var iconName: String?
+    
+}